+++ src/js/store/reducers/index.js
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux';
 import app from './app';
 import errors from './errors';
-import files from './files';
+import scene from './scene';
 import settings from './settings';
 import shield from './shield';
 import user from './user';
@@ -9,7 +9,7 @@ import user from './user';
 const reducers = combineReducers({
     app,
     errors,
-    files,
+    scene,
     settings,
     shield,
     user,
