+++ src/js/store/reducers/scene.js
